+++ CHANGELOG.md
@@ -17,7 +17,7 @@ changes, see each app's individual CHANGELOG:
 - Core `alchemind` interface (see [alchemind changelog](apps/alchemind/CHANGELOG.md#unreleased))
 - OpenAI provider implementation (see [alchemind_openai changelog](apps/alchemind_openai/CHANGELOG.md#unreleased))
 - OpenAI LangChain provider implementation (see [alchemind_openai_langchain changelog](apps/alchemind_openai_langchain/CHANGELOG.md#unreleased))
-  - Support for streaming responses with callback function
+  - Support for streaming responses with callback function via Rust NIF.
 - Provider capabilities table in README
 - Added OpenAI audio transcription (see [alchemind_openai changelog](apps/alchemind_openai/CHANGELOG.md#unreleased))
 - Added OpenAI text-to-speech capabilities (see [alchemind_openai changelog](apps/alchemind_openai/CHANGELOG.md#unreleased))